Add Docker config

+++ script/build-site.sh
@@ -35,8 +35,12 @@ while [[ $# -gt 0 ]]; do
 done
 
 static_build=out
-rm -rf $static_build
-mkdir $static_build
+# Clean output directory (delete contents, not the directory itself for Docker volume mounts)
+if [ -d "$static_build" ]; then
+    rm -rf ${static_build:?}/*
+else
+    mkdir $static_build
+fi
 
 perl mkarchives $static_build
 
@@ -53,7 +57,7 @@ for year in 1999 $(seq 2011 2025); do
         continue
     fi
     target="$static_build/$year"
-    mkdir "$target"
+    mkdir -p "$target"
     cd "$year"
 
     asset_dir="share/static"
